Make the password textarea of KWalletManager monospaced.
I've had trouble recently with the default (proportional) font with the
randomized passwords I try to use, where I was unable to tell apart l
(lowercase L) from I (uppercase i), 0 (zero) from O (the letter), etc.

This is usually much less of a problem with monospaced fonts, so my workaround
is simply to change the default font for the textarea to be monospaced.

If this is a problem just let me know and I'll revert. Likewise if it's felt to
be useful I can backport. (CC'ing kdeutils module coordinator). As of now this
only applies to trunk.

GUI:
CCMAIL:kossebau@kde.org

svn path=/trunk/KDE/kdeutils/kwallet/; revision=1220723