Password contents can now always been shown when navigating the tree
The "show contents" and "hide contenst" buttons now received a menu
with a checkable action letting the user to change the behavior of
the editor when it selects a password entry.

The menu is displayed by long clicking the respective buttons. If
"always show contents" is checked, then the editor will show password
contents right away. No need to click "show contents".

The corresponding setting is put into the rc file so it'll be restored
upon kwalletmanager restart.