Fix XML export regression
StoredTransferJob reads data from the current position of the QIODevice
passed to it. Since a QTemporaryFile is always open in ReadWrite mode,
and we write to the file, the current position is the end of the file.

So we need to rewind the QIODevice's position before passing it to the job.

BUG: 368314
FIXED-IN: 16.08.3