"QFontDatabase::FixedFont" instead of "Monospace"
File src/manager/walletwidget.ui specified font "Monospace" to use
when showing passwords. However, in case there is a font called
"Monospace" installed on the system, this font will be used instead
of the default/system fixed-width font.

If the "Monospace" font is broken or rendered incorrectly, passwords
cannot be read from within the wallet manager.

To fix this problem, this commit adds a line in the corresponding
C++ code in file src/manager/kwalleteditor.cpp which sets a fixed font
by invoking setFont(QFontDatabase::systemFont(QFontDatabase::FixedFont))
which uses the fixed-width font as configured by the user.

Differential Revision: https://phabricator.kde.org/D26288