Use QIcon instead of QPixmap where possible
This avoids having to specify a size in some places.

There is a behavior change. The code clearly intends to fall back to
specific icons when the requested name is not available. However this
does not work since the first call to QIcon::fromTheme will always give
a non-null pixmap since it has a builtin fallback. The new code
correctly makes use of the fallback icons.