Rename option to disable Secret Service to make it more user-friendly
Currently, the option is full of technical jargon and there's no way users can
know what the option does unless they have previous technical knowledge about it
or do a web search for what it means. That's bad from a UX perspective.

As far as end-users are concerned, disabling that option allows third-party
services like KeePassXC to work. So, to make the option more user-friendly,
simplify the option's name and add a description below.

Closes #3