Show window when starting app
Currently when starting the app the window isn't shown unless explicity requested.

This is poor UX since starting the app from Kickoff will result in no window being
shown, which is unexpected and confusing.

The motivation behind this behavior seems to be in that KWallet has an option
to launch the manager as part of the runtime start to show it in the system
tray. In that case it makes sense to not show the window.

However we can better accomodate this use case. kwalletmanager accepts a flag
to indicate tha it is being launched by kwalletd, and kwalletd uses that.

So rework the logic to always show the window on launch/activation, unless
it explicitly comes from kwalletd.

This not only improves the UX, it also somewhat simplifies the code